requirements.txt
brawlstats
python-dotenv
colorama
